+++ src/features/blockchain/types/enums.ts
@@ -1,0 +1,7 @@
+/**
+ * @deprecated use `BlockchainNamespace` instead
+ */
+export enum SupportedBlockchainNamespace {
+  EIP_155 = 'eip155',
+  NEAR = 'near',
+}