+++ src/features/verida/utils/index.ts
@@ -55,11 +55,9 @@ export function getDidClientConfigForNetwork(
  */
 export function getSupportedVeridaNetworks(): EnvironmentType[] {
   const networks: EnvironmentType[] = []
-  if (config.features.veridaMainnet.enabled) {
-    networks.push(EnvironmentType.MAINNET)
-  }
 
-  networks.push(EnvironmentType.TESTNET)
+  networks.push(EnvironmentType.MAINNET)
+  // TODO: Add Banksia when available
 
   if (config.dev.devMode) {
     networks.push(EnvironmentType.DEVNET)
@@ -71,7 +69,5 @@ export function getSupportedVeridaNetworks(): EnvironmentType[] {
 export function getDefaultVeridaNetwork(): EnvironmentType {
   return config.dev.devMode
     ? EnvironmentType.DEVNET // TODO: Should we change to testnet/banksia? As the wallet dev should use a testnet
-    : config.features.veridaMainnet.enabled
-      ? EnvironmentType.MAINNET
-      : EnvironmentType.TESTNET
+    : EnvironmentType.MAINNET
 }