+++ src/pages/Assets/AssetsScreen.tsx
@@ -1,6 +1,6 @@
 import { BottomTabHeaderProps } from '@react-navigation/bottom-tabs'
 import { TabScreenHeader } from 'components'
-import { getSelectedWalletById } from 'features/cryptoWallet'
+import { useSelectedCryptoWallet } from 'features/cryptoWallet'
 import { useThemeAwareStyle } from 'hooks'
 import { Container } from 'native-base'
 import React, { useCallback, useEffect, useMemo, useState } from 'react'
@@ -8,11 +8,10 @@ import { StyleSheet, useWindowDimensions, View } from 'react-native'
 import { SceneMap, TabView } from 'react-native-tab-view'
 
 import { SegmentData, SegmentsControl } from 'components/SegmentControl'
-import WalletNavigationHeader from 'components/WalletSelectorNavigation/WalletNavigationHeader'
+import { WalletNavigationHeader } from 'components/WalletSelectorNavigation/WalletNavigationHeader'
 import WalletSelectorModal from 'components/WalletSelectorNavigation/WalletSelectorModal'
 import { TabsScreenProps } from 'navigation/types'
 import { TokenDashboard } from 'pages/Tokens/TokenDashboard'
-import { useAppSelector } from 'reduxStore/types'
 import { Theme } from 'styles/types'
 
 import Collectibles from './Collectibles'
@@ -49,7 +48,7 @@ type AssetsScreenProps = TabsScreenProps<'Assets'>
 export const AssetsScreen: React.FC<AssetsScreenProps> = (props) => {
   const { navigation } = props
 
-  const selectedWallet = useAppSelector(getSelectedWalletById)
+  const selectedCryptoWallet = useSelectedCryptoWallet()
   const [modalVisible, setModalVisible] = useState(false)
   const [activeTabIndex, setActiveTabIndex] = useState(0)
   const layout = useWindowDimensions()
@@ -67,11 +66,11 @@ export const AssetsScreen: React.FC<AssetsScreenProps> = (props) => {
   const walletSelect = useMemo(
     () => (
       <WalletNavigationHeader
-        selectedWallet={selectedWallet}
-        openWalletModal={openWalletModal}
+        selectedWallet={selectedCryptoWallet}
+        onPress={openWalletModal}
       />
     ),
-    [openWalletModal, selectedWallet]
+    [openWalletModal, selectedCryptoWallet]
   )
 
   useEffect(() => {