+++ CHANGELOG.md
@@ -1,3 +1,34 @@
+# 1.4.1 (2024-05-08)
+
+## Bug fixes
+
+* Fix profile country not updatable
+
+# 1.4.0 (2024-04-26)
+
+## Enhancements
+
+- Improve UI/UX of the crypto wallet list and details
+
+## Bug Fixes
+
+- Fix sending MATIC on Polygon mainnet
+- Fix blockchain cache incorrect expiry strategy
+- Fix selected crypto wallet not kept after restart
+- Fix crash when navigating to "View Seed phrase" of crypto wallet
+
+# 1.3.0 (2024-04-12)
+
+## Enhancements
+
+- Remove Acacia testnet for identity creation and import
+- Handle the Acacia testnet shutdown in the "Not found identity" modal
+
+## Bug fixes
+
+- Fix handling `undefined` asset quote
+- Fix label of coins/tokens in list
+
 # 1.2.2 (2024-03-26)
 
 ## Bug Fixes